# Reduce allocation in InstrumentedChannel