--- a/changelog/@unreleased/pr-733.v2.yml
+++ b/changelog/@unreleased/pr-733.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: '`InstrumentedChannel` now allocates fewer objects per RPC call.'
+  links:
+  - https://github.com/palantir/dialogue/pull/733