## Before this PR

SimulationTest is slow for many reasons. One of my hypothesis is that we actually allocate quite a lot of objects in order to process 45k events.

## After this PR
==COMMIT_MSG==
InstrumentedChannel no longer uses metric-schema's staged builders to access a Timer on every request, instead it does a lookup in a caffeine cache.
==COMMIT_MSG==

This achieved a 20% throughput increase, from 9k -> 11k req/sec

```diff
-com.palantir.dialogue.core.Benchmark - Fired off all requests (4864 ms, 9251req/sec)
+com.palantir.dialogue.core.Benchmark - Fired off all requests (4055 ms, 11097req/sec)
```

![image](https://user-images.githubusercontent.com/3473798/81635544-57fc7200-9409-11ea-9f68-2f00d97753f5.png)



## Possible downsides?
- This isn't JMH precise, so maybe the difference is actually bigger/smaller than my crappy Stopwatch log lines show