@ferozco 2020-05-12
I think there's a trade off between performance and simplicity and I'm worried that the additional complexity of a cache and managing the timing calculations ourselves may not be worth avoiding an allocation