@carterkozak 2020-05-12
I liked the model based on conjure-undertow where we build EndpointChannels which are constructed for a particular endpoint, allowing up-front calculation isolated from caching individual channel components. Thoughts?