@iamdanfox 2020-05-12
Yep I think the per-endpoint build-upfront thing is really what would make life nice here. Just wanted to put this up to show that there is a demonstrable cost (even in crappy benchmarks) to these allocations.