@iamdanfox 2020-05-20
Closing out because we've moved this instrumentation to get bound up-front as part of the EndpointChannel refactor.